TVariant type is now working correctly.

+++ Reflection.h
@@ -8,7 +8,7 @@ typedef struct TVector TVector;
 typedef struct TRtti
 {
     TString Type;
-    uint32_t Size_Of;
+    size_t Size_Of;
 } TRtti;
 
 // Returns runtime type information. 
@@ -19,6 +19,7 @@ TRtti TRtti_Init_(void* check, uint32_t size_Of_Type, TString* type_String);
 // First argument is the type, second argument is optional and also a bool
 // that, if supplied and true, runs a function that returns a large structure.
 #define Rtti(type) &(TRtti){ .Type = *NT_TString(Stringify(type)), .Size_Of = sizeof(type) }
+
 
 void Type_Check(TString* type_String, Array_Of(TRtti) types, uint32_t types_Count);
 bool Compare_Types(TRtti* type_Info_0, TRtti* type_Info_1);
